+++ src/constants.ts
@@ -4,22 +4,34 @@ export const DEFAULT_NETWORK = 'mainnet-beta';
 
 export const TOKEN_MINTS: Record<string, string> = {
   SOL: 'Native',
+  RADR: 'CzFvsLdUazabdiu9TYXujj4EY495fG7VgJJ3vQs6bonk',
   USDC: 'EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v',
   ORE: 'oreoU2P8bN6jkk3jbaiVxYnG1dCXcYxwhwyK9jSybcp',
   BONK: 'DezXAZ8z7PnrnRJjz3wXBoRgixCa6xjnB7YaB1pPB263',
   JIM: 'H9muD33usLGYv1tHvxCVpFwwVSn27x67tBQYH1ANbonk',
   GODL: 'GodL6KZ9uuUoQwELggtVzQkKmU1LfqmDokPibPeDKkhF',
   HUSTLE: 'HUSTLFV3U5Km8u66rMQExh4nLy7unfKHedEXVK1WgSAG',
+  ZEC: 'A7bdiYdS5GjqGFtxf17ppRHtDKPkkRqbKtR27dxvQXaS',
+  CRT: 'CRTx1JouZhzSU6XytsE42UQraoGqiHgxabocVfARTy2s',
+  BLACKCOIN: 'J3rYdme789g1zAysfbH9oP4zjagvfVM2PX7KJgFDpump',
+  GIL: 'CyUgNnKPQLqFcheyGV8wmypnJqojA7NzsdJjTS4nUT2j',
+  ANON: 'D25bi7oHQjqkVrzbfuM6k2gzVNHTSpBLhtakDCzCCDUB',
 };
 
 export const TOKEN_DECIMALS: Record<string, number> = {
   SOL: 9,
+  RADR: 9,
   USDC: 6,
   ORE: 11,
   BONK: 5,
   JIM: 9,
   GODL: 11,
   HUSTLE: 9,
+  ZEC: 8,
+  CRT: 9,
+  BLACKCOIN: 6,
+  GIL: 6,
+  ANON: 9,
 };
 
 export const DEFAULT_PROOF_BIT_LENGTH = 64;